forge fmt

+++ test/MerkleClaim.t.sol
@@ -132,28 +132,28 @@ contract MerkleClaimTest is Test {
         }
     }
 
-    function testTransferAssetsOut(address[] memory assets) public{
+    function testTransferAssetsOut(address[] memory assets) public {
         // example ERC20 to etch our fake ERC20 addresses at
         address example = address(new EXAMPLEERC20());
 
-        uint[] memory amounts = new uint[](assets.length);
-        for(uint i; i < amounts.length; ++i){
+        uint256[] memory amounts = new uint256[](assets.length);
+        for (uint256 i; i < amounts.length; ++i) {
             // bound address by left shifting to not be a precompile
             assets[i] = address(bytes20(assets[i]) << 8);
             // bound amounts as the uint representation of the addresses to get random numbers without needing to generate another address array and bound its size
-            amounts[i] = bound(uint(bytes32(bytes20(assets[i]))), 1, 100 ether);
+            amounts[i] = bound(uint256(bytes32(bytes20(assets[i]))), 1, 100 ether);
             // if first time with this address, etch it
-            if(assets[i].code.length == 0){
+            if (assets[i].code.length == 0) {
                 vm.etch(assets[i], example.code);
             }
             // deal tokens to it. If an address is used multiple times, deal multiple times
-            uint prebal = ERC20(assets[i]).balanceOf(address(claim));
-            deal(assets[i], address(claim), (100 ether)+prebal);
+            uint256 prebal = ERC20(assets[i]).balanceOf(address(claim));
+            deal(assets[i], address(claim), (100 ether) + prebal);
         }
 
         // transfer the assets
         claim.transferAssets(assets, amounts, address(this));
-        for(uint i; i < amounts.length; ++i){
+        for (uint256 i; i < amounts.length; ++i) {
             // assert not that these amounts are equal but rather that they modulo to zero...
             // This is because an address can be delt multiple times, with the same token amount
             // That's why this is also valid if 2x or 3x the amounts are received as it's withdrawn in 2 instances in the transferAssets arrays
